PXE: Pass root device hints via kernel cmdline
The way we pass information from Ironic to the PXE deploy ramdisk is via
the kernel cmdline, this patch is adding a new 'root_device' argument to
the PXE template which will contain the root device hints so the ramdisk
knows which device it should pick to deploy the image onto. If no hints
are specified then 'root_device' option won't be passed.

Change-Id: I527f3d3f3115dbfd08d3a6f76eeb14712f29adca
Implements: blueprint root-device-hints

Conflicts:
	ironic/tests/drivers/test_iscsi_deploy.py