PXE/ISCSI: Call partprobe after partioning the disk
The _wait_for_disk_to_become_available function from
common/disk_partitioner.py calls fuser to check if the is not being held
by another process, but it doesn't actually check if the partitions of the
device is available after the disk is partitioned. This patch adds a call
to partprobe to make the kernel re-read the partition table of the device
and then list the partitions to verify if they are really available.

Closes-Bug: #1403887
Change-Id: I880ef1120f98e32d802f52db36955da95f7daeb7

Conflicts:
	ironic/tests/drivers/test_deploy_utils.py